doctoto2/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-crispy-forms==2.1
django-formtools==2.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
Faker==20.0.3
inflection==0.5.1
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
numpy==1.26.0
packaging==23.1
pandas==2.1.0
plotly==5.17.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.12.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
